backend/requirements.txt
# Lightweight app packages — fast layer.

# Core API
fastapi[standard]>=0.115.0
pydantic>=2.7
pydantic-settings>=2.3
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.48
asyncpg>=0.29.0
alembic>=1.14.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0,<5.0
python-multipart>=0.0.9

# Cache / Queue client
redis>=5.0.0

# Rate limiting
slowapi>=0.1.9

# Logging
structlog>=24.4.0

# Storage
boto3>=1.34.0

# Dev / test
ruff>=0.5.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
factory-boy>=3.3.0
